# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Los Angeles Crime Wire API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SubscriptionInputEditionType = typeof SubscriptionInputEditionType[keyof typeof SubscriptionInputEditionType];


export const SubscriptionInputEditionType = {
  digital: 'digital',
  mailed: 'mailed',
  both: 'both',
} as const;

export interface SubscriptionInput {
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  zip?: string | null;
  editionType: SubscriptionInputEditionType;
  consent: boolean;
}

export interface Subscription {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  zip?: string | null;
  editionType: string;
  createdAt: string;
}

export interface TipInput {
  /** @nullable */
  nameOrAlias?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  message: string;
  /** @nullable */
  provenance?: string | null;
}

export interface Tip {
  id: number;
  /** @nullable */
  nameOrAlias?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  message: string;
  /** @nullable */
  provenance?: string | null;
  createdAt: string;
}

export type ListSubscriptionsParams = {
format?: ListSubscriptionsFormat;
};

export type ListSubscriptionsFormat = typeof ListSubscriptionsFormat[keyof typeof ListSubscriptionsFormat];


export const ListSubscriptionsFormat = {
  json: 'json',
  csv: 'csv',
} as const;