# lib/api-zod/src/generated/types/subscription.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Los Angeles Crime Wire API
 * OpenAPI spec version: 0.1.0
 */

export interface Subscription {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  zip?: string | null;
  editionType: string;
  createdAt: string;
}