# lib/api-zod/src/generated/types/subscriptionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Los Angeles Crime Wire API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionInputEditionType } from './subscriptionInputEditionType';

export interface SubscriptionInput {
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  zip?: string | null;
  editionType: SubscriptionInputEditionType;
  consent: boolean;
}