# lib/api-zod/src/generated/types/tip.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Los Angeles Crime Wire API
 * OpenAPI spec version: 0.1.0
 */

export interface Tip {
  id: number;
  /** @nullable */
  nameOrAlias?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  message: string;
  /** @nullable */
  provenance?: string | null;
  createdAt: string;
}