# lib/api-zod/src/generated/types/tipInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Los Angeles Crime Wire API
 * OpenAPI spec version: 0.1.0
 */

export interface TipInput {
  /** @nullable */
  nameOrAlias?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  message: string;
  /** @nullable */
  provenance?: string | null;
}